更新cli文件结构
