+++ httpfpt/__init__.py
@@ -2,5 +2,6 @@
 # -*- coding: utf-8 -*-
 from httpfpt.core.get_conf import set_project_config as set_project_config
 from httpfpt.core.path_conf import set_project_dir as set_project_dir
+from httpfpt.run import run as run
 
 __version__ = 'v0.5.1'
