+++ httpfpt/utils/enum_control.py
@@ -1,6 +1,11 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
-from enum import Enum
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from enum import Enum
 
 
 def get_enum_values(enum_class: type[Enum]) -> list:
